Create best engineering college in delhi ncr
As technological advancement is taking place constantly at a very high pace, so in order to make the students compatible to deal with these advancements, there must be special attention given to the technical improvement of the students. With an aim to embed a technology oriented education to the students, Geeta Engineering College offers a 4 yrs B-Tech Degree in the field of Computer Science &Engineering, Electronics & Communication Engineering, Civil Engineering and Mechanical Engineering in Panipat, Haryana after the student has passed class 12th. B-Tech can also be selected as the course for the students who have cleared Diploma and can directly get admission to second year of B-Tech in the disciplinary area of their polytechnic through lateral entry. Geeta Engineering College aims at nurturing the budding talents and imbibing in them the ability to exceed and bring out their best. GEC is known for its Best Performance in Academics and has even been recognized as the Top Engineering Colleges in the region of Haryana and Delhi-NCR. Moreover the glory of the institute has reached its height after the achievement of Academic Excellence Award in the field of imparting Technical education to the students.

The best placement record, the presence of highly qualified faculty members who are expertise in their field of interest, the brilliancy in students and the tendency to learn and bring out some new and productive output has helped Geeta Engineering College to reach the peak of Success. The students of GEC are excellent in their Academic, social, moral and technical ethics. Join the best B.Tech courses here at Geeta Group, we are one of the top class engineering college in North India with the best placement opportunity

B.Tech Courses In Haryana
S.NO.	COURSES	SEATS	DURATION
1.	 Mechanical Engineering	120	4 Yrs.
2.	 Computer Science Engineering	60	4 Yrs.
3.	 Electronics and Communication Engineering	60	4 Yrs.
4.	 Civil Engineering	60